# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Structure of Iglu's Shop API functions",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/outlet": {
            "get": {
                "tags": [
                    "Outlet"
                ],
                "summary": "Get all products with a discount",
                "description": "Returns a list of all the products with discount.\n\nThis is a public endpoint .\n\nThis function retrieves all products with a discount from the\nProduct model and returns them in JSON format. Each product includes\nid, name, description, price, stock, image, has_discount, discount, provider_id, category_id, created_at, and updated_at.\nid, name, description, price, stock, image,has_discount,discount,provider_id,category_id,created_at,updated_at.",
                "operationId": "955d73daad4a3e784296234cb13fc5d8",
                "responses": {
                    "200": {
                        "description": "List of products with discount (public endpoint, does not require authentication).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Outlet"
                ],
                "summary": "Creates a new product from outlet",
                "description": "Creates a new product from outlet.\n\nRequires authentication and administrator permissions.\n\nWhen creating a new product, the following fields are required:\n\n- `name`: The name of the product. Must be a string with a minimum of 3 characters.\n- `description`: A detailed description of the product. Must be a string with a maximum of 255 characters.\n- `price`: The price of the product. Must be a decimal number with up to two decimal places.\n- `category_id`: ID of the category the product belongs to. Must be an integer and must exist in the database.\n- `provider_id`: ID of the provider associated with the product. Must be an integer and must exist in the database.\n\nThe following fields are optional, but if included, must meet the specified validations:\n\n- `stock`: An integer indicating the available quantity of the product in stock.\n- `image`: A valid URL pointing to the product's image.\n- `has_discount`: A boolean field indicating whether the product has a discount. This field must always be true for outlet products.\n   To create a product without a discount, use the appropriate products endpoint.\n- `discount`: The discount applied to the product. If `has_discount` is true, this value must be greater than 0 and less than 100.\n\nValidation includes type constraints and specific value rules. If any field does not comply with the rules, an error will be returned.",
                "operationId": "d11b65ac7954125f05487fe45862fcff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OutletCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product successfully created."
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 21
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/outlet/{id}": {
            "get": {
                "tags": [
                    "Outlet"
                ],
                "summary": "Get an outlet product by ID",
                "description": "Returns the data of a specific outlet product.\n\nThis is a public endpoint.\n\nIt searches for the product by its ID and checks if it belongs to the outlet.\nIf not found, it returns an error.",
                "operationId": "7029b3370902ee0537345e1fa4f65ce5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found or not belonging to the outlet",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Outlet"
                ],
                "summary": "Delete an outlet product",
                "description": "Deletes a product from the outlet by its ID.\n\nRequires authentication and administrator permissions.\n\nIf the product does not exist, it returns a 404 error.\nIf successfully deleted, it returns a 200 status code.",
                "operationId": "97899162908852e82c2655aedb3de42b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product successfully deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NotFoundError": {
                "title": "Resource not found",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found"
                    }
                },
                "type": "object"
            },
...
            "OutletCreate": {
                "required": [
                    "name",
                    "description",
                    "price",
                    "category_id",
                    "provider_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 3,
                        "example": "Sports T-shirt"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "High-performance technical t-shirt"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 100
                    },
                    "image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/product.jpg"
                    },
                    "has_discount": {
                        "description": "Must be true or equivalent (accepted). The created product must have a discount",
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "example": true
                    },
                    "discount": {
                        "description": "Must be a value greater than 0 and less than or equal to 100.",
                        "type": "number",
                        "example": 15
                    },
                    "category_id": {
                        "description": "Related category ID (foreign key to Category)",
                        "type": "integer",
                        "example": 1
                    },
                    "provider_id": {
                        "description": "Related provider ID (foreign key to Provider)",
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "title": "Product",
                "required": [
                    "id",
                    "name",
                    "description",
                    "price",
                    "stock",
                    "image",
                    "has_discount",
                    "discount",
                    "provider_id",
                    "category_id",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sports T-shirt"
                    },
                    "description": {
                        "type": "string",
                        "example": "High-performance technical t-shirt"
                    },
                    "price": {
                        "type": "number",
                        "example": 29.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 2
                    },
                    "image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://www.example.com/images/tshirt.jpg"
                    },
                    "has_discount": {
                        "description": "Este campo debe ser siempre false. si se cambia a true el producto pasa al outlet",
                        "type": "boolean",
                        "default": 0,
                        "example": false
                    },
                    "discount": {
                        "description": "Si has_discount es falsy entonces discount debe ser 0",
                        "type": "number",
                        "example": 0
                    },
                    "provider_id": {
                        "description": "Related provider ID (foreign key a Provider)",
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "description": "Related category ID (foreign key a Category)",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-05 21:45:59"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-05 21:45:59"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "title": "Error de validación",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "error": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Name is required"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}